# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Academy Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Approve / reject / block / unblock user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reset-device:
    post:
      operationId: resetUserDevice
      tags: [users]
      summary: Reset device lock for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device reset

  # ── Admins (superadmin only) ───────────────────────────────────────────────
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Upload note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Certificates ──────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List all sent notifications (admin view)
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/my:
    get:
      operationId: getMyNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: My notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read

  # ── Activity Logs ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        mobileNumber:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        profilePhoto:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        interestedCourse:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
        createdAt:
          type: string
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        mobileNumber:
          type: string
        profilePhoto:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        reason:
          type: ["string", "null"]
...
    Note:
      type: object
      required: [id, fileName, courseId, driveFileId, uploadedAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        driveFileId:
          type: string
        driveViewUrl:
          type: ["string", "null"]
        facultyId:
          type: ["integer", "null"]
        facultyName:
          type: ["string", "null"]
        uploadedAt:
          type: string
    NoteInput:
      type: object
      required: [fileName, courseId, driveFileId]
      properties:
        fileName:
          type: string
        courseId:
          type: integer
        driveFileId:
          type: string
        driveViewUrl:
          type: ["string", "null"]

    # ── Certificate schemas ────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, title, message, target, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        imageUrl:
          type: ["string", "null"]
        target:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    NotificationInput:
      type: object
      required: [title, message, target]
      properties:
        title:
          type: string
        message:
          type: string
        imageUrl:
          type: ["string", "null"]
        target:
          type: string

    # ── Activity Log schemas ───────────────────────────────────────────────
...